+++ R/functions.R
@@ -75,7 +75,7 @@
 #'  forwardsolve(l, x, k = ncol(l), upper.tri = FALSE,
 #'               transpose = FALSE)
 #'
-#'  #'  # miscellaneous operations
+#'  # miscellaneous operations
 #'  aperm(x, perm)
 #'  apply(x, MARGIN, FUN = c("sum", "max", "mean", "min",
 #'                           "prod", "cumsum", "cumprod"))
@@ -102,7 +102,7 @@
 #'   \code{cospi()}, \code{sinpi()}, and \code{tanpi()} do not use the
 #'   computationally more stable routines to compute \code{cos(x * pi)} etc.
 #'   that are available in R under some operating systems. Similarly
-#'   \code{trigamma()} uses tensorflows polygamma function, resulting in lower
+#'   \code{trigamma()} uses TensorFlow's polygamma function, resulting in lower
 #'   precision than R's equivalent.
 #'
 #' @examples